+++ index.html
@@ -12,6 +12,8 @@
     <link href="https://fonts.googleapis.com/css?family=Open+Sans&display=swap" rel="stylesheet">
     <script type="module" src="/js/physi.js"></script>
     <script type="module" src="/js/three.js"></script>
+    <script type="module" src="/scripts/audio.js"></script>
+
 
   </head>
   <body>
