+++ scripts/audio.js
@@ -1,0 +1,23 @@
+      function sound(src) {
+        this.sound = document.createElement("audio");
+        this.sound.src = src;
+        this.sound.setAttribute("preload", "auto");
+        this.sound.setAttribute("controls", "none");
+        this.sound.style.display = "none";
+        document.body.appendChild(this.sound);
+        this.play = function(){
+          this.sound.play();
+        }
+        this.stop = function(){
+          this.sound.pause();
+        }
+        this.loop = function(){
+          this.sound.loop = true;
+          this.sound.play();
+        }
+        
+        return this
+      }
+
+export default {sound};
+
